upload avatart has been implemented
