+++ src/controllers/profile.js
@@ -24,7 +24,10 @@ exports.updateProfile = async (req, res, next) => {
   const user = await User.findOneAndUpdate(
     { _id: req.user._id, type: req.user.type },
     {
-      $set: req.body,
+      $set: {
+        avatar: req.files ? req.files[0].path : req.body.avatar,
+        ...req.body,
+      },
     },
     { new: true }
   );
