+++ src/routes/profile.js
@@ -4,6 +4,7 @@ const { validate } = require('../middlewares/bodyValidator');
 const { isAuth } = require('../middlewares/auth');
 const { validateProfile } = require('../middlewares/validatorSchemas');
 const profileController = require('../controllers/profile');
+const uploadHandler = require('../middlewares/google-upload');
 
 router.get('/', isAuth, profileController.getProfile);
 router.patch(
@@ -11,6 +12,7 @@ router.patch(
   isAuth,
   validateProfile,
   validate,
+  uploadHandler.any('avatar'),
   profileController.updateProfile
 );
